Changed the way that subviews receive appear/disappear events.
Now, the default action of willAppear and
willDisappear is just to notify subviews
that they are about to appear/disappear.
If a ViewHandler overrides willAppear or
willDisappear, then it is responsible for notifying
its subviews of appear and disappear events. 2 new
methods, subviewsWillAppear and subviewsWillDisappear
were created to make this easier for view handlers
that override these methods.

Related to #7