First pass at deleting people.
The person object now extends the Deletable object,
which includes a bit indicating deleted status, the
date that the object was deleted, and the reason that
the object was deleted.

Currently only people can be deleted, and there are no tools
for viewing deleted people or 'resurrecting' them. Additionally,
terminology around this action needs to be standardized.