Fixed bug where doses would be shown when there were no vaccines.
Formerly, after the last vaccine was removed, its doses would
still be displayed to the right. This commit fixes that and makes
the vaccine administration panel behave similarly to the reminder
administration panel in that it doesn't display anything but an
empty list when there are no vaccines in the system.

Fixes #33