First pass at vaccine appointment rescheduling.
Now, when a vaccine appointment is rescheduled or the dose is
administered, the rest of the doses in the series are inspected
and rescheduled if they fall within any 'minimum interval'
windows.

It should be noted that the code currently only works for changes
forward in time, and does not properly reschedule the appts
backwards in time.