Changed the way reminder times are set.
Previously, the time of day that a reminder should be sent would
be entered into a text field that wasn't checked for validity.

Now, users are asked to select a time of day from a combobox that
contains only times on the hour. The ReminderDispatcher thread
awakes hourly and sends the scheduled reminders.

Fixes #34