First pass at a flag listener
The new class, FlagListener waits for a form submission
and then evaluates it against its list of flags.
If any flags are triggered, then a ui alert
is shown to the user.

In the future, an SMS message will be sent out, but
for now a UI alert is used for testing purposes.